requirements.txt
kivy
buildozer
pillow
cython
